USer Points dataType update

+++ IdentityServer/IdenitityServer.Core/Common/Interfaces/IUserService.cs
@@ -23,6 +23,6 @@ namespace IdenitityServer.Core.Common.Interfaces
         Task<UserClaimModel> AddUserClaimsBasedOnUserId(UserClaimModel userClaimModel);
         Task<UserClaimModel> ModifyUserClaimsBasedOnUserId(UserClaimModel userClaimModel);
         Task<string> GetUserIdByVendorClaim(string vendorId);
-        Task<bool> UpdateUserPoints(string userId, int amountDebited);
+        Task<bool> UpdateUserPoints(string userId, double amountDebited);
     }
 }
